+++ ctx/context.go
@@ -5,8 +5,9 @@ package ctx
 // Context are a cumulative structure carried over to each module,
 // to contain data later steps might require
 type Context struct {
+	Artifacts   Artifacts
 	ProjectName string
+	Publish     bool
 	TargetDir   string
 	Version     string
-	Artifacts   Artifacts
 }